Local Makefiles for frontend and backend
- Move existing Makefile to `frontend/`, and update `FRONTEND` var to
`.`.
- Update `nunjucksbuild.js` to take the template path as an argument to
make it compatible with the move by taking `NUNJUCKS_TEMPLATES` as an
argument.
- Add a stub Makefile to `backend/`, with a `.PHONY` `all` target, that
just runs a nop to make `make` happy. This file will be fleshed out as
we begin working on the backend, or if no compilation/preprocessing is
needed, will stay as-is (or be removed entirely, but I prefer keeping
it personally).
- Add a new super-Makefile to the top level that just runs sub-makes
in the frontend and backend directories.